+++ cases/testcases/unused_fn_args/multimethods/green.clj
@@ -1,0 +1,6 @@
+(ns testcases.unused-fn-args.multimethods.green)
+
+(defmulti foo
+  (fn [a b c]
+    ;; all args unintentionally unused
+    (rand)))