+++ cases/testcases/unused_fn_args/multimethods/red.clj
@@ -1,0 +1,6 @@
+(ns testcases.unused-fn-args.multimethods.red)
+
+(defn foo []
+  (fn [a b c]
+    ;; all args unintentionally unused
+    (rand)))