+++ src/eastwood/util.clj
@@ -938,10 +938,14 @@ of these kind."
 (defn process-configs [warning-enable-config]
   (reduce (fn [configs {:keys [linter] :as m}]
             (case linter
-              (:constant-test :redefd-vars :unused-ret-vals
-                              :wrong-tag :suspicious-test
-                              :unused-meta-on-macro
-                              :unused-ret-vals-in-try)
+              (:constant-test
+               :redefd-vars
+               :suspicious-test
+               :unused-fn-args
+               :unused-meta-on-macro
+               :unused-ret-vals
+               :unused-ret-vals-in-try
+               :wrong-tag)
               (update-in configs [linter]
                          conj (dissoc m :linter))
               :suspicious-expression
@@ -1045,7 +1049,13 @@ of these kind."
            w linter (format " for invocation of macro '%s'" macro-symbol)
            suppress-conditions opt)))
 
-      (:unused-ret-vals :unused-ret-vals-in-try :constant-test :wrong-tag :suspicious-test :unused-meta-on-macro)
+      (:constant-test
+       :suspicious-test
+       :unused-fn-args
+       :unused-meta-on-macro
+       :unused-ret-vals
+       :unused-ret-vals-in-try
+       :wrong-tag)
       (let [suppress-conditions (get-in opt [:warning-enable-config linter])]
         (allow-warning-based-on-enclosing-macros
          w linter "" suppress-conditions opt)))))